Revert "Arreglo de css"
