--- public/CatalogoAccesoriosE.css
@@ -1,241 +1,0 @@
-:root {
-    --primary: #a51314;        /* Rojo principal */
-    --primary-dark: #701106;   /* Rojo oscuro */
-    --dark: #050506;           /* Negro */
-    --dark-gray: #1a1a1a;      /* Gris oscuro */
-    --medium-gray: #333333;    /* Gris medio */
-    --light: #f7f7f7;          /* Blanco */
-}
-
-body {
-    background-color: #f8f9fa;
-    color: #212529;
-    padding-left: 0; /* Eliminamos padding por defecto */
-  }
-  
-  .sidebar {
-    background-color: var(--dark) !important;
-    position: fixed;
-    top: 0;
-    left: 0;
-    height: 100vh;
-    overflow-y: auto;
-    z-index: 1000;
-    width: 250px; /* Ancho fijo del sidebar */
-    padding-top: 15px;
-  }
-  
-  .nav-link {
-    color: var(--light) !important;
-    padding: 0.75rem 1rem;
-    border-radius: 0.25rem;
-    margin: 0.25rem 0;
-    transition: all 0.3s ease;
-  }
-  
-  .nav-link.active {
-    background-color: var(--primary) !important;
-    color: white !important;
-  }
-  
-  .nav-link:hover:not(.active) {
-    background-color: var(--medium-gray);
-  }
-  
-  main {
-    margin-left: 250px; /* Mismo ancho que el sidebar */
-    padding: 20px;
-    padding-top:0;
-    width: calc(100% - 250px); /* Ajustamos el ancho restando el sidebar */
-  }
-  
-
-/* Ajuste para pantallas pequeñas */
-@media (max-width: 991.98px) {
-    .sidebar {
-        width: 100%;
-        height: auto;
-        position: relative;
-    }
-    
-    main {
-        margin-left: 0;
-        width: 100%;
-    }
-}
-
-.stats-card {
-    background-color: white;
-    border: none;
-    border-radius: 10px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    transition: transform 0.3s;
-}
-
-.stats-card:hover {
-    transform: translateY(-5px);
-}
-
-.stats-card h6 {
-    color: #6c757d;
-    margin-bottom: 8px; /* Espacio adecuado */
-}
-
-.dashboard-card {
-    background-color: white;
-    border: none;
-    border-radius: 10px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    transition: transform 0.3s;
-    height: 100%; /* Para mantener altura uniforme */
-}
-
-.dashboard-card:hover {
-    transform: translateY(-5px);
-}
-
-.filter-section {
-    background-color: white;
-    border-radius: 10px;
-    padding: 20px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-    margin-bottom: 25px; /* Espacio consistente */
-}
-
-/* Estilo consistente para botones */
-.btn-primary {
-    background-color: var(--primary);
-    border-color: var(--primary);
-    transition: all 0.3s ease;
-}
-
-.btn-primary:hover {
-    background-color: var(--primary-dark);
-    border-color: var(--primary-dark);
-}
-
-.btn-outline-primary {
-    color: var(--primary);
-    border-color: var(--primary);
-    transition: all 0.3s ease;
-}
-
-.btn-outline-primary:hover {
-    background-color: var(--primary);
-    color: white;
-    border-color: var(--primary);
-}
-
-/* Aseguramos que los botones secundarios sean consistentes */
-.btn-outline-secondary {
-    transition: all 0.3s ease;
-}
-
-.text-primary {
-    color: var(--primary) !important;
-}
-
-/* Corregimos estilos para bordes */
-.border-bottom {
-    border-bottom: 1px solid #dee2e6 !important;
-    margin-bottom: 20px; /* Espacio adecuado */
-}
-
-.modal-content {
-    border-radius: 10px;
-    overflow: hidden; /* Para que el header redondeado se vea bien */
-}
-
-.inventory-badge {
-    font-size: 0.8rem;
-    padding: 5px 10px;
-    border-radius: 4px;
-}
-
-/* Ajustamos la estructura de las tarjetas */
-.card-body {
-    padding: 1.25rem;
-}
-
-.card-footer {
-    background-color: transparent;
-    padding: 1rem 1.25rem;
-    border-top: 1px solid rgba(0, 0, 0, 0.125);
-}
-
-/* Mejoramos el perfil de usuario */
-.user-profile {
-    padding: 15px;
-    border-top: 1px solid rgba(255, 255, 255, 0.1);
-}
-
-/* Aseguramos que los filtros tengan suficiente espacio */
-.form-label {
-    margin-bottom: 0.5rem;
-    font-weight: 500;
-}
-
-/* Mejoramos los modales */
-.modal-header.bg-primary {
-    color: white;
-}
-
-/* Estilos específicos para la tabla de accesorios */
-#accesoriosTable img.img-thumbnail {
-    width: 80px;
-    height: 80px;
-    object-fit: cover;
-    border-radius: 5px;
-}
-
-#accesoriosTable .badge {
-    font-size: 0.85rem;
-    padding: 6px 10px;
-    border-radius: 4px;
-}
-
-/* Estilo para la imagen placeholder */
-.bg-secondary {
-    background-color: #6c757d !important;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 80px;
-    height: 80px;
-    border-radius: 5px;
-}
-
-/* Estilo para modales de detalle de accesorios */
-.modal-lg .img-fluid {
-    max-height: 250px;
-    object-fit: contain;
-}
-
-.list-group-item.bg-transparent {
-    background-color: transparent !important;
-    border-color: rgba(255, 255, 255, 0.125);
-}
-
-/* Estilos para DataTables */
-.dataTables_wrapper .dataTables_paginate .paginate_button.current {
-    background: var(--primary) !important;
-    color: white !important;
-    border-color: var(--primary) !important;
-}
-
-.dataTables_wrapper .dataTables_paginate .paginate_button:hover {
-    background: var(--primary-dark) !important;
-    color: white !important;
-    border-color: var(--primary-dark) !important;
-}
-
-.dataTables_wrapper .dataTables_length select,
-.dataTables_wrapper .dataTables_filter input {
-    border: 1px solid #ced4da;
-    border-radius: 4px;
-    padding: 0.375rem 0.75rem;
-}
-
-.dataTables_wrapper .dataTables_info {
-    padding-top: 10px;
-}
