+++ public/admin.css
@@ -1,0 +1,346 @@
+:root {
+    --primary: #a51314;         /* Rojo principal */
+    --primary-dark: #701106;    /* Rojo oscuro para hover y elementos secundarios */
+    --primary-light: #c82333;   /* Rojo claro para hover */
+    --dark: #050506;            /* Negro para fondos oscuros */
+    --dark-gray: #1a1a1a;       /* Gris oscuro para tarjetas */
+    --medium-gray: #333333;     /* Gris medio para bordes */
+    --light-gray: rgba(255,255,255,0.05); /* Gris claro para filas alternativas */
+    --light: #f7f7f7;           /* Casi blanco para texto claro */
+}
+
+body {
+    font-family: 'Montserrat', sans-serif;
+    background-color: #000000;
+    min-height: 100vh;
+    display: flex;
+    margin: 0;
+    color: var(--light);
+    padding-left: 250px; /* Space for sidebar */
+}
+
+/* Header y Navegación */
+.site-header {
+    background-color: #6c0b09f4;
+    width: 250px;
+    position: fixed;
+    left: 0;
+    top: 0;
+    bottom: 0;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 1rem 0;
+    box-shadow: 2px 0 10px rgba(0, 0, 0, 0.2);
+    z-index: 1000;
+    transition: width 0.3s ease;
+}
+
+.logo-container {
+    display: flex;
+    align-items: center;
+    margin-bottom: 2rem;
+}
+
+.logo-img {
+    width: 160px;
+    height: 140px;
+    max-height: 200px;
+}
+
+.navbar-nav {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.navbar-nav .nav-item {
+    width: 90%;
+    margin: 0.5rem 0;
+}
+
+.navbar-nav .nav-link {
+    color: var(--light);
+    font-weight: 600;
+    padding: 0.75rem 1rem;
+    width: 100%;
+    border-radius: 8px;
+    display: flex;
+    align-items: center;
+    gap: 0.75rem;
+    transition: all 0.3s ease;
+}
+
+.navbar-nav .nav-link i {
+    margin-right: 0.5rem;
+}
+
+.navbar-nav .nav-link:hover {
+    background-color: var(--primary-dark);
+    color: var(--light);
+}
+
+
+.navbar-nav .nav-link.active::after {
+    width: 80%;
+}
+
+@media (max-width: 992px) {
+    .site-header {
+        background-color: var(--dark);
+    }
+
+    .navbar-collapse {
+        background-color: var(--dark);
+        position: fixed;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100vh;
+        padding: 20px;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        opacity: 0;
+        visibility: hidden;
+        transition: opacity 0.3s ease, visibility 0.3s ease;
+    }
+
+    .navbar-collapse.show {
+        opacity: 1;
+        visibility: visible;
+    }
+
+    .navbar-nav {
+        flex-direction: column;
+        align-items: center;
+    }
+
+    .navbar-nav .nav-item {
+        margin: 1rem 0;
+    }
+
+    .navbar-nav .nav-link {
+        font-size: 1.5rem;
+    }
+}
+
+/* Scroll behavior */
+@media (min-width: 992px) {
+    .site-header.scrolled {
+        background-color: rgba(5, 5, 6, 0.95);
+        box-shadow: 0 2px 15px rgba(0,0,0,0.2);
+    }
+}
+
+header {
+    background-color: #6c0b09f4;
+    text-align: center;
+    padding: 20px;
+    width: 100%;
+    box-shadow: 0 2px 10px rgba(0,0,0,0.2);
+}
+
+header h1 {
+    color: var(--light);
+    margin: 0;
+    font-weight: 700;
+}
+
+main {
+    flex: 1;
+    text-align: center;
+    padding: 20px;
+    background-color: var(--dark);
+}
+
+ul {
+    list-style: none;
+    padding: 0;
+    display: flex;
+    justify-content: center;
+    flex-wrap: wrap;
+    gap: 20px;
+    margin-bottom: 2rem;
+}
+
+ul li a {
+    color: var(--light);
+    text-decoration: none;
+    padding: 0.5rem 1rem;
+    border-radius: 30px;
+    transition: all 0.3s ease;
+    font-weight: 600;
+    position: relative;
+}
+
+ul li a:hover {
+    color: var(--primary);
+    background-color: rgba(255,255,255,0.1);
+}
+
+ul li a:hover::after {
+    content: '';
+    position: absolute;
+    width: 80%;
+    height: 3px;
+    background-color: var(--primary);
+    bottom: -5px;
+    left: 50%;
+    transform: translateX(-50%);
+}
+
+footer {
+    margin-top: auto;
+    text-align: center;
+    padding: 20px;
+    background-color: var(--dark-gray);
+    color: var(--light);
+}
+
+footer ul {
+    margin-top: 1rem;
+}
+
+footer ul li a {
+    color: var(--light-gray);
+    text-decoration: none;
+    transition: color 0.3s ease;
+}
+
+footer ul li a:hover {
+    color: var(--primary);
+}
+
+.chart-container {
+    position: relative;
+    width: 100%;
+    max-width: 800px;
+    height: 400px; /* Fixed height */
+    margin: 20px auto;
+    background-color: rgb(255, 255, 255);
+    border-radius: 10px;
+    padding: 1rem;
+}
+
+canvas {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100% !important;
+    height: 100% !important;
+    max-width: 100%;
+    max-height: 100%;
+}
+
+button {
+    background-color: var(--primary);
+    color: var(--light);
+    border: none;
+    padding: 0.75rem 2rem;
+    font-size: 16px;
+    margin-top: 20px;
+    border-radius: 30px;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    font-weight: 600;
+    text-transform: uppercase;
+}
+
+button:hover {
+    background-color: var(--primary-light);
+    transform: translateY(-5px);
+    box-shadow: 0 10px 20px rgba(0,0,0,0.2);
+}
+
+@media (max-width: 600px) {
+    ul {
+        flex-direction: column;
+        align-items: center;
+    }
+
+    header h1 {
+        font-size: 1.5rem;
+    }
+
+    button {
+        width: 90%;
+        padding: 0.5rem 1rem;
+    }
+
+    .chart-container {
+        height: 250px; /* Adjusted height for mobile */
+    }
+}
+
+/* Footer */
+footer {
+    background-color: var(--secondary);
+    color: white;
+    padding: 3rem 0 1rem;
+}
+
+.footer-brand {
+    font-family: 'Racing Sans One', cursive;
+    color: var(--primary);
+    font-size: 1.5rem;
+    margin-bottom: 1rem;
+}
+
+.footer-links h5 {
+    color: white;
+    font-weight: 600;
+    margin-bottom: 1.5rem;
+}
+
+.footer-links ul {
+    list-style: none;
+    padding: 0;
+}
+
+.footer-links li {
+    margin-bottom: 0.75rem;
+}
+
+.footer-links a {
+    color: rgba(255,255,255,0.7);
+    text-decoration: none;
+    transition: all 0.3s ease;
+}
+
+.footer-links a:hover {
+    color: var(--primary);
+    padding-left: 5px;
+}
+
+.social-links {
+    display: flex;
+    gap: 15px;
+}
+
+.social-links a {
+    width: 40px;
+    height: 40px;
+    border-radius: 50%;
+    background-color: rgba(255,255,255,0.1);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    color: white;
+    font-size: 1.25rem;
+    transition: all 0.3s ease;
+}
+
+.social-links a:hover {
+    background-color: var(--primary);
+    transform: translateY(-5px);
+}
+
+.copyright {
+    padding-top: 2rem;
+    margin-top: 2rem;
+    border-top: 1px solid rgba(255,255,255,0.1);
+    text-align: center;
+}
