--- public/administrador.css
@@ -1,369 +1,0 @@
-/* administrador.css */
-:root {
-  --primary: #a51314;
-  --primary-dark: #701106;
-  --primary-light: #f1d6d6;
-  --dark: #050506;
-  --light: #f7f7f7;
-  --gray: #333333;
-  --light-gray: #e9ecef;
-  --success: #198754;
-  --success-light: #d1e7dd;
-  --danger: #dc3545;
-  --danger-light: #f8d7da;
-  --warning: #ffc107;
-  --info: #0dcaf0;
-  --purple: #9d4edd;
-}
-
-  
-  body {
-    font-family: 'Montserrat', sans-serif;
-    background-color: var(--light);
-    display: flex;
-    min-height: 100vh;
-    margin: 0;
-  }
-  
-  /* Sidebar */
-  .sidebar {
-    width: 280px;
-    background-color: var(--dark);
-    color: white;
-    position: fixed;
-    height: 100vh;
-    overflow-y: auto;
-    z-index: 1000;
-    box-shadow: 2px 0 10px rgba(0, 0, 0, 0.1);
-    transition: all 0.3s ease;
-  }
-  
-.sidebar-header {
-  padding: 1.5rem;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-  text-align: center;
-}
-
-.sidebar-brand img {
-  height: 120px;
-  object-fit: contain;
-  transition: transform 0.3s ease;
-}
-
-.sidebar-brand img:hover {
-  transform: scale(1.05);
-}
-
-.sidebar-nav {
-  padding: 1rem 0;
-}
-
-.user-profile {
-  padding: 1rem 1.5rem;
-  margin-bottom: 1rem;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-  display: flex;
-  align-items: center;
-}
-
-.user-avatar {
-  width: 45px;
-  height: 45px;
-  border-radius: 50%;
-  overflow: hidden;
-  margin-right: 12px;
-  border: 2px solid var(--primary);
-}
-
-.user-avatar img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.user-info {
-  flex: 1;
-}
-  
-
-.user-name {
-  font-weight: 600;
-  margin: 0;
-  color: white;
-  font-size: 0.95rem;
-}
-
-.user-role {
-  color: rgba(255, 255, 255, 0.7);
-  font-size: 0.8rem;
-  margin: 0;
-}
-
-.nav-link {
-  color: rgba(255, 255, 255, 0.8);
-  padding: 0.9rem 1.5rem;
-  display: flex;
-  align-items: center;
-  transition: all 0.3s;
-  border-left: 4px solid transparent;
-  border-radius: 0 30px 30px 0;
-  margin: 0.25rem 0;
-}
-
-.nav-link:hover,
-.nav-link.active {
-  background-color: var(--primary) !important;
-  color: white;
-  border-left: 4px solid var(--primary);
-}
-
-.nav-link i {
-  margin-right: 12px;
-  font-size: 1.25rem;
-}
-
-  /* Main Content */
-  .main-content {
-    margin-left: 280px;
-    padding: 2rem;
-    flex: 1;
-    transition: all 0.3s ease;
-  }
-  
-  .content-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 2rem;
-    padding-bottom: 1rem;
-    border-bottom: 1px solid var(--light-gray);
-  }
-  
-  .content-header h1 {
-    font-weight: 700;
-    color: var(--gray);
-    font-size: 1.8rem;
-    margin: 0;
-  }
-  
-  /* Stats Cards */
-  .stats-row {
-    display: flex;
-    gap: 1rem;
-    margin-bottom: 2rem;
-  }
-  
-  .stat-card {
-    flex: 1;
-    background: white;
-    border-radius: 12px;
-    padding: 1.5rem;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
-    display: flex;
-    align-items: center;
-    position: relative;
-    overflow: hidden;
-    transition: transform 0.3s ease, box-shadow 0.3s ease;
-  }
-  
-  .stat-card:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 10px 20px rgba(0, 0, 0, 0.1);
-  }
-  
-  .stat-card.bg-primary {
-    background-color: var(--primary) !important;
-    color: white;
-  }
-  
-  .stat-card.bg-success {
-    background-color: var(--success) !important;
-    color: white;
-  }
-  
-  .stat-card.bg-warning {
-    background-color: var(--warning) !important;
-    color: var(--dark);
-  }
-  
-  .stat-card.bg-info {
-    background-color: var(--info) !important;
-    color: white;
-  }
-  
-  .stat-card.bg-purple {
-    background-color: var(--purple) !important;
-    color: white;
-  }
-  
-  .card-icon {
-    font-size: 2.5rem;
-    opacity: 0.2;
-    position: absolute;
-    right: 1.5rem;
-    top: 50%;
-    transform: translateY(-50%);
-  }
-  
-  /* Cards */
-  .card {
-    border: none;
-    border-radius: 12px;
-    box-shadow: 0 6px 15px rgba(0, 0, 0, 0.07);
-    transition: all 0.3s ease;
-    margin-bottom: 1.5rem;
-  }
-  
-  .card:hover {
-    transform: translateY(-5px);
-    box-shadow: 0 12px 20px rgba(0, 0, 0, 0.1);
-  }
-  
-  .card-header {
-    background-color: white;
-    border-bottom: 1px solid var(--light-gray);
-    font-weight: 600;
-    padding: 1.25rem 1.5rem;
-    border-radius: 12px 12px 0 0 !important;
-  }
-  
-  .card-body {
-    padding: 1.5rem;
-  }
-  
-  /* Charts */
-  .chart-container {
-    position: relative;
-    height: 300px;
-    width: 100%;
-  }
-  
-  .chart-card {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-  }
-  
-  /* Tables */
-  .table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-bottom: 0;
-  }
-  
-  .table th {
-    background-color: var(--light-gray);
-    color: var(--gray);
-    font-weight: 600;
-    padding: 0.75rem;
-    text-align: left;
-  }
-  
-  .table td {
-    padding: 0.75rem;
-    border-bottom: 1px solid var(--light-gray);
-    vertical-align: middle;
-  }
-  
-  .table-hover tbody tr:hover {
-    background-color: rgba(0, 0, 0, 0.02);
-  }
-  
-  /* Badges */
-  .badge {
-    padding: 0.35em 0.65em;
-    font-weight: 600;
-  }
-  
-  .badge-primary {
-    background-color: var(--primary);
-  }
-  
-  .badge-success {
-    background-color: var(--success);
-  }
-  
-  .badge-warning {
-    background-color: var(--warning);
-    color: var(--dark);
-  }
-  
-  .badge-danger {
-    background-color: var(--danger);
-  }
-  
-  /* Responsive */
-  @media (max-width: 992px) {
-    .sidebar {
-      width: 240px;
-    }
-    .main-content {
-      margin-left: 240px;
-    }
-  }
-  
-  @media (max-width: 768px) {
-    .sidebar {
-      width: 100%;
-      height: auto;
-      position: relative;
-    }
-    .main-content {
-      margin-left: 0;
-      padding: 1rem;
-    }
-    .stats-row {
-      flex-direction: column;
-    }
-  }
-  
-  /* Tooltips */
-  .tooltip {
-    position: relative;
-    display: inline-block;
-  }
-  
-  .tooltip .tooltiptext {
-    visibility: hidden;
-    width: 120px;
-    background-color: var(--dark);
-    color: #fff;
-    text-align: center;
-    border-radius: 6px;
-    padding: 5px;
-    position: absolute;
-    z-index: 1;
-    bottom: 125%;
-    left: 50%;
-    margin-left: -60px;
-    opacity: 0;
-    transition: opacity 0.3s;
-  }
-  
-  .tooltip:hover .tooltiptext {
-    visibility: visible;
-    opacity: 1;
-  }
-  
-  /* Utility Classes */
-  .text-end {
-    text-align: right;
-  }
-  
-  .text-uppercase {
-    text-transform: uppercase;
-  }
-  
-  .mb-0 {
-    margin-bottom: 0 !important;
-  }
-  
-  /* Animation */
-  @keyframes fadeIn {
-    from { opacity: 0; transform: translateY(20px); }
-    to { opacity: 1; transform: translateY(0); }
-  }
-  
-  .animate-fade {
-    animation: fadeIn 0.5s ease-out forwards;
-  }
