+++ public/catalogo.css
@@ -21,16 +21,14 @@ body {
 
 /* Header y Navegación */
 .site-header {
-    background-color: #000000b0; /* Más transparencia para mejor visibilidad */
+    background-color: rgba(0, 0, 0, 0.664); /* Más transparencia para mejor visibilidad */
     position: fixed;
     width: 100%;
     top: 0;
     z-index: 1000;
     box-shadow: 0 2px 10px rgba(0,0,0,0.2); /* Sombra más suave */
     transition: background-color 0.3s ease;
-    font-size: 17px;
 }
-
 
 .navbar {
     padding: 0.5rem 0;
@@ -58,6 +56,7 @@ body {
 
 .navbar-nav {
     align-items: center;
+    margin-right: -9rem;
 }
 
 .navbar-nav .nav-item {
@@ -66,7 +65,7 @@ body {
 
 .navbar-nav .nav-link {
     color: var(--light);
-    font-weight: 600;
+    font-weight: 800;
     padding: 0.5rem 1rem;
     position: relative;
     transition: all 0.3s ease;
