+++ public/empleado.css
@@ -1,0 +1,429 @@
+:root {
+    --primary: #a51314;         /* Rojo principal */
+    --primary-dark: #701106;    /* Rojo oscuro para hover y elementos secundarios */
+    --primary-light: #c82333;   /* Rojo claro para hover */
+    --dark: #050506;            /* Negro para fondos oscuros */
+    --dark-gray: #1a1a1a;       /* Gris oscuro para tarjetas */
+    --medium-gray: #333333;     /* Gris medio para bordes */
+    --light-gray: rgba(255,255,255,0.05); /* Gris claro para filas alternativas */
+    --light: #f7f7f7;           /* Casi blanco para texto claro */
+}
+
+body {
+    font-family: 'Montserrat', sans-serif;
+    background-color: #000;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    overflow-x: hidden;
+    scroll-behavior: smooth;
+}
+
+/* Header y Navegación */
+.site-header {
+    background-color: rgba(0, 0, 0, 0.8); /* Más opaco */
+    position: fixed;
+    width: 100%;
+    top: 0;
+    z-index: 1000;
+    box-shadow: 0 2px 10px rgba(0, 0, 0, 0.3);
+    transition: background-color 0.3s ease;
+}
+
+.navbar {
+    padding: 0.5rem 1rem;
+}
+
+.logo-container {
+    display: flex;
+    align-items: center;
+}
+
+.logo-img {
+    width: 160px;
+    height: auto;
+    max-height: 200px;
+}
+
+.navbar-toggler {
+    border: none;
+    color: var(--light);
+}
+
+.navbar-nav {
+    align-items: center;
+    margin-right: -9rem;
+}
+
+.navbar-nav .nav-item {
+    margin: 0 0.5rem;
+}
+
+.navbar-nav .nav-link {
+    color: var(--light);
+    font-weight: bold;
+    padding: 0.5rem 1rem;
+    position: relative;
+    transition: all 0.3s ease;
+}
+
+.navbar-nav .nav-link:hover {
+    color: var(--primary);
+}
+
+.navbar-nav .nav-link::after {
+    content: '';
+    position: absolute;
+    width: 0;
+    height: 3px;
+    background-color: var(--primary);
+    bottom: -5px;
+    left: 50%;
+    transform: translateX(-50%);
+    transition: width 0.3s ease;
+}
+
+.navbar-nav .nav-link:hover::after {
+    width: 80%;
+}
+
+.navbar-nav .nav-link.active {
+    color: var(--primary);
+}
+
+.navbar-nav .nav-link.active::after {
+    width: 80%;
+}
+
+/* Responsive adjustments */
+@media (max-width: 992px) {
+    .site-header {
+        background-color: var(--dark);
+    }
+
+    .navbar-collapse {
+        background-color: var(--dark);
+        position: fixed;
+        top: 0;
+        left: 0;
+        width: 100%;
+        height: 100vh;
+        padding: 20px;
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        opacity: 0;
+        visibility: hidden;
+        transition: opacity 0.3s ease, visibility 0.3s ease;
+    }
+
+    .navbar-collapse.show {
+        opacity: 1;
+        visibility: visible;
+    }
+
+    .navbar-nav {
+        flex-direction: column;
+        align-items: center;
+    }
+}
+
+/* Hero Section */
+.hero {
+    background-image: url('https://example.com/your-image.jpg');
+    background-size: cover;
+    background-position: center;
+    background-attachment: fixed;
+    min-height: 85vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    color: white;
+    padding: 3rem 1rem;
+    position: relative;
+}
+
+.hero::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: rgba(0, 0, 0, 0.6);
+    z-index: 1;
+}
+
+.hero .container {
+    position: relative;
+    z-index: 2;
+}
+
+.hero h1 {
+    font-size: 4rem;
+    font-weight: bold;
+    margin-bottom: 1rem;
+}
+
+.hero p {
+    font-size: 1.75rem;
+    margin-bottom: 1.5rem;
+}
+
+.brand-buttons {
+    display: flex;
+    gap: 10px;
+}
+
+.brand-buttons button {
+    padding: 10px 20px;
+    border-radius: 10px;
+    transition: background-color 0.3s, transform 0.3s;
+}
+
+.brand-buttons button:hover {
+    background-color: var(--primary-dark);
+    transform: scale(1.05);
+}
+
+/* Motorcycle Grid */
+.motorcycle-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+    gap: 30px;
+    padding: 20px;
+}
+
+.motorcycle-card {
+    background: #fff;
+    border-radius: 10px;
+    overflow: hidden;
+    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease;
+}
+
+.motorcycle-card:hover {
+    transform: translateY(-10px);
+}
+
+.motorcycle-card img {
+    width: 100%;
+    height: 200px;
+    object-fit: cover;
+    transition: transform 0.3s ease;
+}
+
+.motorcycle-card:hover img {
+    transform: scale(1.05);
+}
+
+.motorcycle-details {
+    padding: 20px;
+}
+
+.motorcycle-details h2 {
+    color: var(--dark);
+    font-size: 1.5rem;
+    margin-bottom: 10px;
+}
+
+.motorcycle-details h2::after {
+    content: '';
+    display: block;
+    width: 50px;
+    height: 3px;
+    background: var(--primary);
+    margin: 10px 0;
+}
+
+.price {
+    color: var(--primary);
+    font-size: 1.6rem;
+    font-weight: bold;
+    margin: 15px 0;
+}
+
+/* Modal Styles */
+.modal-content {
+    background: #1a1a1a;
+    color: white;
+}
+
+.spec-item {
+    display: flex;
+    justify-content: space-between;
+    padding: 12px 0;
+    border-bottom: 1px solid rgba(255, 255, 255, 0.1);
+}
+
+.spec-label {
+    color: rgba(255, 255, 255, 0.7);
+}
+
+.spec-value {
+    color: white;
+}
+
+/* Footer */
+footer {
+    background-color: var(--dark-gray);
+    color: white;
+    padding: 3rem 0;
+}
+
+.footer-brand {
+    color: var(--primary);
+    font-size: 1.5rem;
+    margin-bottom: 1rem;
+}
+
+.footer-links h5 {
+    color: white;
+    margin-bottom: 1.5rem;
+}
+
+.footer-links ul {
+    list-style: none;
+    padding: 0;
+}
+
+.footer-links li {
+    margin-bottom: 0.75rem;
+}
+
+.footer-links a {
+    color: rgba(255,255,255,0.7);
+    text-decoration: none;
+}
+
+.footer-links a:hover {
+    color: var(--primary);
+}
+
+/* Animations */
+@keyframes fadeInUp {
+    from {
+        opacity: 0;
+        transform: translateY(20px);
+    }
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
+}
+
+/* Estilos adicionales para mejor responsive design */
+@media (max-width: 576px) {
+    .navbar-brand {
+        font-size: 1.2rem;
+    }
+    
+    .card-img-top {
+        height: 180px;
+        object-fit: cover;
+    }
+    
+    .btn {
+        padding: 0.375rem 0.5rem;
+        font-size: 0.875rem;
+    }
+    
+    h3 {
+        font-size: 1.5rem;
+        text-align: center;
+    }
+    
+    .footer-brand {
+        font-size: 1.5rem;
+    }
+}
+
+/* Mejoras para tablets */
+@media (min-width: 577px) and (max-width: 991px) {
+    .card-img-top {
+        height: 200px;
+        object-fit: cover;
+    }
+}
+
+/* Estilos generales para mejoras responsive */
+.card {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+.card-body {
+    flex: 1 1 auto;
+    display: flex;
+    flex-direction: column;
+}
+
+.card-text {
+    margin-bottom: 0.5rem;
+}
+
+.card-actions {
+    margin-top: auto;
+    display: flex;
+    gap: 5px;
+}
+
+@media (max-width: 767px) {
+    .card-actions {
+        flex-direction: column;
+    }
+    
+    .card-actions .btn {
+        width: 100%;
+        margin-bottom: 5px;
+    }
+}
+
+/* Footer responsive */
+footer {
+    background-color: #212529;
+    color: #fff;
+    padding: 2rem 0 1rem;
+}
+
+.footer-links ul {
+    padding-left: 0;
+    list-style: none;
+}
+
+.footer-links ul li {
+    margin-bottom: 0.5rem;
+}
+
+.footer-links ul li a {
+    color: #aaa;
+    text-decoration: none;
+    transition: color 0.3s;
+}
+
+.footer-links ul li a:hover {
+    color: #fff;
+}
+
+.social-links a {
+    display: inline-block;
+    margin-right: 1rem;
+    color: #fff;
+    font-size: 1.2rem;
+}
+
+.copyright {
+    margin-top: 2rem;
+    text-align: center;
+    font-size: 0.875rem;
+    color: #aaa;
+}
+
+@media (max-width: 767px) {
+    footer h5 {
+        margin-top: 1.5rem;
+    }
+}
