--- public/gestionE.css
@@ -1,431 +1,0 @@
-:root {
-  --primary: #a51314;
-  --primary-dark: #701106;
-  --primary-light: #f1d6d6;
-  --dark: #050506;
-  --light: #f7f7f7;
-  --gray: #333333;
-  --light-gray: #e9ecef;
-  --success: #198754;
-  --success-light: #d1e7dd;
-  --danger: #dc3545;
-  --danger-light: #f8d7da;
-  --warning: #ffc107;
-  --info: #0dcaf0;
-  --purple: #9d4edd;
-}
-
-body {
-  font-family: 'Montserrat', sans-serif;
-  background-color: var(--light);
-  display: flex;
-  min-height: 100vh;
-  margin: 0;
-}
-
-/* Sidebar Styles */
-.sidebar {
-  width: 280px;
-  background-color: var(--dark);
-  color: white;
-  position: fixed;
-  height: 100vh;
-  overflow-y: auto;
-  z-index: 1000;
-  box-shadow: 2px 0 10px rgba(0, 0, 0, 0.1);
-  transition: all 0.3s ease;
-}
-
-.sidebar-header {
-  padding: 1.5rem;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-  text-align: center;
-}
-
-.sidebar-brand img {
-  height: 120px;
-  object-fit: contain;
-  transition: transform 0.3s ease;
-}
-
-.sidebar-brand img:hover {
-  transform: scale(1.05);
-}
-
-.sidebar-nav {
-  padding: 1rem 0;
-}
-
-.user-profile {
-  padding: 1rem 1.5rem;
-  margin-bottom: 1rem;
-  border-bottom: 1px solid rgba(255, 255, 255, 0.1);
-  display: flex;
-  align-items: center;
-}
-
-.user-avatar {
-  width: 45px;
-  height: 45px;
-  border-radius: 50%;
-  overflow: hidden;
-  margin-right: 12px;
-  border: 2px solid var(--primary);
-}
-
-.user-avatar img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.user-info {
-  flex: 1;
-}
-
-.user-name {
-  font-weight: 600;
-  margin: 0;
-  color: white;
-  font-size: 0.95rem;
-}
-
-.user-role {
-  color: rgba(255, 255, 255, 0.7);
-  font-size: 0.8rem;
-  margin: 0;
-}
-
-.nav-link {
-  color: rgba(255, 255, 255, 0.8);
-  padding: 0.9rem 1.5rem;
-  display: flex;
-  align-items: center;
-  transition: all 0.3s;
-  border-left: 4px solid transparent;
-  border-radius: 0 30px 30px 0;
-  margin: 0.25rem 0;
-}
-
-.nav-link:hover,
-.nav-link.active {
-  background-color: var(--primary) !important;
-  color: white;
-  border-left: 4px solid var(--primary);
-}
-
-.nav-link i {
-  margin-right: 12px;
-  font-size: 1.25rem;
-}
-
-/* Main Content Styles */
-.main-content {
-  margin-left: 280px;
-  padding: 2rem;
-  flex: 1;
-  transition: all 0.3s ease;
-}
-
-.content-header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 2rem;
-  padding-bottom: 1rem;
-  border-bottom: 1px solid var(--light-gray);
-}
-
-.content-header h1 {
-  font-weight: 700;
-  color: var(--gray);
-  font-size: 1.8rem;
-  margin: 0;
-}
-
-.breadcrumbs {
-  color: var(--gray);
-  font-size: 0.9rem;
-}
-
-/* Employee Cards */
-.employee-card {
-  background-color: white;
-  border-radius: 16px;
-  box-shadow: 0 6px 15px rgba(0, 0, 0, 0.07);
-  overflow: hidden;
-  margin-bottom: 2rem;
-  transition: all 0.3s ease;
-  border: 1px solid transparent;
-}
-
-.employee-card:hover {
-  transform: translateY(-5px);
-  box-shadow: 0 12px 20px rgba(0, 0, 0, 0.1);
-  border-color: var(--primary-light);
-}
-
-.employee-img {
-  height: 240px;
-  background-size: cover;
-  background-position: center;
-  position: relative;
-  border-bottom: 3px solid var(--primary);
-}
-
-.employee-status {
-  position: absolute;
-  bottom: 15px;
-  right: 15px;
-  padding: 0.3rem 0.8rem;
-  border-radius: 20px;
-  font-size: 0.8rem;
-  font-weight: bold;
-  box-shadow: 0 3px 6px rgba(0, 0, 0, 0.1);
-}
-
-.status-active {
-  background-color: var(--success);
-  color: white;
-}
-
-.status-inactive {
-  background-color: var(--danger);
-  color: white;
-}
-
-.employee-info {
-  padding: 1.5rem;
-}
-
-.employee-name {
-  font-size: 1.3rem;
-  font-weight: 600;
-  color: var(--primary-dark);
-  margin-bottom: 0.8rem;
-  border-bottom: 1px dashed var(--light-gray);
-  padding-bottom: 0.5rem;
-}
-
-.employee-detail {
-  font-size: 0.95rem;
-  color: var(--gray);
-  margin-bottom: 0.5rem;
-  display: flex;
-  justify-content: space-between;
-}
-
-.employee-detail strong {
-  color: var(--primary-dark);
-  min-width: 100px;
-}
-
-.employee-actions {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 1rem;
-  padding-top: 1rem;
-  border-top: 1px solid var(--light-gray);
-}
-
-.btn-edit {
-  color: #fff;
-  background-color: #6c757d;
-  border: none;
-  padding: 0.375rem 0.75rem;
-  border-radius: 0.25rem;
-  transition: all 0.2s;
-}
-
-.btn-edit:hover {
-  background-color: #5a6268;
-  transform: translateY(-2px);
-}
-
-.btn-fire {
-  color: white;
-  background-color: var(--danger);
-  border: none;
-  padding: 0.375rem 0.75rem;
-  border-radius: 0.25rem;
-  transition: all 0.2s;
-}
-
-.btn-rehire {
-  color: white;
-  background-color: var(--success);
-  border: none;
-  padding: 0.375rem 0.75rem;
-  border-radius: 0.25rem;
-  transition: all 0.2s;
-}
-
-.btn-fire:hover,
-.btn-rehire:hover,
-.btn-edit:hover {
-  opacity: 0.9;
-  transform: translateY(-2px);
-}
-
-/* Dashboard Stats */
-.stats-row {
-  display: flex;
-  gap: 1rem;
-  margin-bottom: 2rem;
-}
-
-.stat-card {
-  flex: 1;
-  background: white;
-  border-radius: 12px;
-  padding: 1.5rem;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
-  display: flex;
-  align-items: center;
-  overflow: hidden;
-  position: relative;
-}
-
-.stat-icon {
-  background-color: var(--primary-light);
-  color: var(--primary);
-  width: 60px;
-  height: 60px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 12px;
-  margin-right: 1rem;
-  font-size: 1.8rem;
-}
-
-.active-employees .stat-icon {
-  background-color: var(--success-light);
-  color: var(--success);
-}
-
-.inactive-employees .stat-icon {
-  background-color: var(--danger-light);
-  color: var(--danger);
-}
-
-.stat-info h3 {
-  font-size: 1.8rem;
-  font-weight: 700;
-  margin: 0;
-  color: var(--gray);
-}
-
-.stat-info p {
-  margin: 0;
-  color: var(--gray);
-  font-size: 0.9rem;
-}
-
-/* Top action buttons */
-.action-buttons .btn {
-  padding: 0.5rem 1rem;
-  border-radius: 8px;
-  font-weight: 500;
-  transition: all 0.2s;
-  margin-left: 0.5rem;
-}
-
-.action-buttons .btn:hover {
-  transform: translateY(-2px);
-}
-
-.action-buttons .btn-primary {
-  background-color: var(--primary);
-  border-color: var(--primary);
-}
-
-.action-buttons .btn-primary:hover {
-  background-color: var(--primary-dark);
-}
-
-.action-buttons .btn-dark {
-  background-color: var(--dark);
-}
-
-/* Search bar */
-.search-container {
-  position: relative;
-  max-width: 400px;
-  margin-bottom: 2rem;
-}
-
-.search-input {
-  width: 100%;
-  padding: 0.75rem 1rem 0.75rem 3rem;
-  border-radius: 50px;
-  border: 1px solid var(--light-gray);
-  transition: all 0.3s;
-}
-
-.search-input:focus {
-  outline: none;
-  box-shadow: 0 0 0 3px rgba(165, 19, 20, 0.2);
-  border-color: var(--primary);
-}
-
-.search-icon {
-  position: absolute;
-  left: 1rem;
-  top: 50%;
-  transform: translateY(-50%);
-  color: var(--gray);
-}
-
-/* Responsive */
-@media (max-width: 992px) {
-  .sidebar {
-    width: 240px;
-  }
-  .main-content {
-    margin-left: 240px;
-  }
-}
-
-@media (max-width: 768px) {
-  .sidebar {
-    width: 100%;
-    height: auto;
-    position: relative;
-  }
-  .main-content {
-    margin-left: 0;
-    padding: 1rem;
-  }
-  .stats-row {
-    flex-direction: column;
-  }
-}
-
-/* Tooltip */
-.tooltip {
-  position: relative;
-  display: inline-block;
-}
-
-.tooltip .tooltiptext {
-  visibility: hidden;
-  width: 120px;
-  background-color: var(--dark);
-  color: #fff;
-  text-align: center;
-  border-radius: 6px;
-  padding: 5px;
-  position: absolute;
-  z-index: 1;
-  bottom: 125%;
-  left: 50%;
-  margin-left: -60px;
-  opacity: 0;
-  transition: opacity 0.3s;
-}
-
-.tooltip:hover .tooltiptext {
-  visibility: visible;
-  opacity: 1;
-}
